+++ src/lib/scriptengine/rsscript_test.rs
@@ -47,7 +47,7 @@ fn get_script_runner_cargo_other() {
     let provider = get_script_runner();
     envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
 
-    assert_eq!(provider, ScriptRunner::CargoScript);
+    assert_eq!(provider, ScriptRunner::RustScript);
 }
 
 #[test]
@@ -56,11 +56,13 @@ fn execute_default_valid() {
     if test::should_test(false) {
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
 
-        execute(
+        let valid = execute(
             &vec!["fn main() {println!(\"test\");}".to_string()],
+            None,
             &vec![],
             true,
         );
+        assert!(valid);
     }
 }
 
@@ -73,6 +75,7 @@ fn execute_default_not_compile() {
 
         execute(
             &vec!["fn main() {donotcompile();}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -90,6 +93,7 @@ fn execute_default_runtime_panic() {
 
         execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -104,11 +108,13 @@ fn execute_default_runtime_panic_no_validate() {
     if test::should_test(false) {
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
 
-        execute(
+        let valid = execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             false,
         );
+        assert!(!valid);
 
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
     }
@@ -120,11 +126,13 @@ fn execute_rust_script_valid() {
     if test::should_test(false) {
         envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
 
-        execute(
+        let valid = execute(
             &vec!["fn main() {println!(\"test\");}".to_string()],
+            None,
             &vec![],
             true,
         );
+        assert!(valid);
 
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
     }
@@ -139,6 +147,7 @@ fn execute_rust_script_not_compile() {
 
         execute(
             &vec!["fn main() {donotcompile();}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -156,6 +165,7 @@ fn execute_rust_script_runtime_panic() {
 
         execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -170,11 +180,13 @@ fn execute_rust_script_runtime_panic_no_validate() {
     if test::should_test(false) {
         envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
 
-        execute(
+        let valid = execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             false,
         );
+        assert!(!valid);
 
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
     }
@@ -189,6 +201,7 @@ fn execute_cargo_play_not_compile() {
 
         execute(
             &vec!["fn main() {donotcompile();}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -206,6 +219,7 @@ fn execute_cargo_play_runtime_panic() {
 
         execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             true,
         );
@@ -220,11 +234,13 @@ fn execute_cargo_play_runtime_panic_no_validate() {
     if test::should_test(false) {
         envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "cargo-play");
 
-        execute(
+        let valid = execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],
+            None,
             &vec![],
             false,
         );
+        assert!(!valid);
 
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
     }