+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGELOG
 
+### v0.37.8
+
+* Enhancement: Support script runners and shebang in condition scripts #987
+* Enhancement: Support script runner args for rust scripts #1025
+* Enhancement: Support condition_script as string and not just vec (similar to script)
+
 ### v0.37.7 (2024-01-14)
 
 * Fix: Detect crate installation fix for new cargo list format